+++ authority/provisioner/oidc_test.go
@@ -323,7 +323,7 @@ func TestOIDC_AuthorizeSign(t *testing.T) {
 				assert.Equals(t, sc.StatusCode(), tt.code)
 				assert.Nil(t, got)
 			} else if assert.NotNil(t, got) {
-				assert.Equals(t, 7, len(got))
+				assert.Equals(t, 8, len(got))
 				for _, o := range got {
 					switch v := o.(type) {
 					case *OIDC:
@@ -343,6 +343,8 @@ func TestOIDC_AuthorizeSign(t *testing.T) {
 						assert.Equals(t, string(v), "name@smallstep.com")
 					case *x509NamePolicyValidator:
 						assert.Equals(t, nil, v.policyEngine)
+					case *WebhookController:
+						assert.Len(t, 0, v.webhooks)
 					default:
 						assert.FatalError(t, fmt.Errorf("unexpected sign option of type %T", v))
 					}